requirements.txt
# EHR System Requirements

# Web Framework
flask>=3.0.0
werkzeug>=3.0.0

# Audio Processing
sounddevice>=0.5.0
soundfile>=0.13.0
numpy>=2.0.0

# Speech Recognition (Whisper) - OPTIONAL: only needed if you want a local fallback when Gemini API is unavailable
# openai-whisper>=20231117

# AI/ML for EHR Extraction & Transcription (Gemini)
google-generativeai>=0.3.0

# Hybrid Recommendation Engine
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Note: Full LIME/SHAP can be added later for production
lime>=0.2.0
# shap>=0.42.0

# Federated Learning Framework
flwr>=1.5.0

# PyTorch for Whisper fine-tuning
torch>=2.0.0
torchaudio>=2.0.0

# Data handling
datasets>=2.14.0

# Audio processing for training
librosa>=0.10.0

# Metrics calculation
jiwer>=3.0.0

# Environment variables
python-dotenv>=1.0.0
